% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrimpact-package.R
\docType{package}
\name{centrimpact-package}
\alias{centrimpact}
\alias{centrimpact-package}
\title{centrimpact: Supporting the Analysis and Visualization of Community Engaged Research Metrics}
\description{
The `centrimpact` R package provides a comprehensive toolkit for analyzing and visualizing community engaged research metrics as highlighted in the CEnTR*IMPACT framework. It offers specialized functions for researcher-partner alignment, project dynamics, cascade effects, and creating publication-ready visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CENTRIMPACT/centrimpact-review}
  \item Report bugs at \url{https://github.com/CENTRIMPACT/centrimpact-review/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Forest Price \email{jfprice@iu.edu} (\href{https://orcid.org/0000-0002-6506-3526}{ORCID})

}
\keyword{internal}
